+++ src/Classes/Variables/Float.php
@@ -1,0 +1,67 @@
+<?php
+/**
+ * Copyright © 2020 Quorrax, S. L.
+ *
+ * This file is part of quorrax/variable.
+ *
+ * quorrax/variable is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * quorrax/variable is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with quorrax/variable. If not, see <https://www.gnu.org/licenses/>.
+ */
+
+namespace Quorrax\Classes\Variables;
+
+use InvalidArgumentException;
+use Quorrax\Classes\Variable;
+use Quorrax\Interfaces\Variables\Float as FloatInterface;
+
+/**
+ * @package Quorrax\Classes\Variables
+ */
+class Float extends Variable implements FloatInterface
+{
+    /**
+     * @var float
+     */
+    private $value;
+
+    /**
+     * @return float
+     */
+    public function getValue()
+    {
+        return $this->value;
+    }
+
+    /**
+     * @param float $value
+     *
+     * @return \Quorrax\Interfaces\Variables\Float
+     */
+    public function setValue($value)
+    {
+        if (is_float($value)) {
+            $this->value = $value;
+        } else {
+            throw new InvalidArgumentException("The given argument for the {\$value} parameter is not a float.");
+        }
+        return $this;
+    }
+
+    /**
+     * @param float $value
+     */
+    public function __construct($value = 0.0)
+    {
+        $this->setValue($value);
+    }
+}